chore(tests): Remove unnecessary roadrunner dependencies
- Removed spiral/roadrunner-cli and spiral/roadrunner-http from the composer dependencies.
- This cleanup helps streamline the dependency management process.

+++ app/Http/Middleware/VerifySignature.php
@@ -62,7 +62,7 @@ final class VerifySignature
             'nonce' => ['required', 'string', 'size:16'],
             'timestamp' => \sprintf(
                 'required|int|max:%s|min:%s',
-                $time = (Date::now()->timestamp + 1),
+                $time = ((int) Date::now()->timestamp + 1),
                 $time - $effectiveTime
             ),
         ])->validate();
